homework_21/requirements.txt
Django>=3.2
 djangorestframework
 pytest
 pytest-django
